Set up the Node structure and made the debug properly functional
Added a node structure to the bvh menu and set up the debug menu using that node structure. Currently the Node is only used to find the level of each box, but we can use it to find the intersection if we use it like a tree structure.
Also currently the debug only shows the chosen layer of boxes, it is easy to switch it to show everything leading up to that point by changing the == to >= in debugdraw

+++ src/main.cpp
@@ -89,7 +89,7 @@ float softShadow(glm::vec3 intersectionPos, SphericalLight sphericalLight, const
         if (castShadow(intersectionPos, center, bvh)) return 1;              // If in shadow skip further computation for this lightsource
     }
     // we generate random points as point lights
-    int contributions = 0;
+    float contributions = 0.0f;
     for (glm::vec3 rpoint : randomVectors) {
         if (castShadow(intersectionPos, center + rpoint * r, bvh)) contributions++;
     }
